requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
google-genai>=1.0.0
gTTS==2.4.0
requests==2.31.0
Pillow
Werkzeug==3.0.1